==> R_scripts/5_sub-clustering_mixed_cluster.Rmd <==
---
title: "sub-clustering of cluster id #4"
author: "Yuvarani Masarapu"
date: "11/12/2021"
output: html_document
---

```{r eval=FALSE}
#We sub-clustered the mixed cluster here to investigate it further at a more granular level
```

## Loading package libraries
```{r}
library(SeuratObject)
library(Seurat)
library(ggplot2)
```

## Read data
```{r}
lung.integrated <- readRDS("~/data.object.integrated.rds")

# Subset for the 'Mixed' cluster
mixed <- subset(lung.integrated, idents = 4)
```

#### Sub-clustering
```{r}
set.seed(63)
mixed  <- FindNeighbors(object = mixed, assay = "SCT", dims = 1:35, k.param = 23, graph.name = "graph_afterHarmony", reduction = "harmony_sid")

mixed <- FindClusters(object = mixed, pc.use = 1:35, resolution = 0.15, save.SNN = T, do.sparse = T, graph.name = "graph_afterHarmony", random.seed = 3915, group.singletons = TRUE)
```

Custom colors
```{r eval=FALSE}
library(RColorBrewer)
n <- 60
qual_col_pals = brewer.pal.info[brewer.pal.info$category == 'qual',]

col_vector = unlist(mapply(brewer.pal, qual_col_pals$maxcolors, rownames(qual_col_pals)))
```

```{r fig.height=5, fig.width=7}
names(col_vector) = (levels(mixed$seurat_clusters))

DimPlot(object = mixed, group.by = "seurat_clusters", reduction = "umap_after_harmony", label = T, pt.size = 0.2, cols = col_vector[1:3]) + ggtitle("Seurat clusters distribution in UMAP")

DimPlot(object = mixed, reduction = "umap_after_harmony", split.by = "sample_type", cols = col_vector[1:3]) + ggtitle("Visualise two sample types side-by-side")
```

### DE analysis
```{r eval=FALSE}
markers.list <- lapply(levels(mixed$seurat_clusters), function(cl){
  cl <- FindMarkers(object = mixed, assay = "SCT", ident.1 = cl, ident.2 = NULL, test.use = "wilcox", random.seed = 47, min.pct = 0.005)
})

for (list in 1:length(markers.list)){
  markers.list[[list]] = cbind(gene = rownames(markers.list[[list]]), markers.list[[list]])
}

saveRDS(markers.list, file = "DEgenes_subclusters_cluster4_mixed.rds")
writexl::write_xlsx(markers.list, path = "mixed_subcluster_markers.xlsx", col_names = TRUE, format_headers = TRUE)
```

```{r}
saveRDS(mixed, file = "mixed_final_subclusters.rds")
```

```{r}
p = SpatialPlot(mixed, group.by = "seurat_clusters", cols = col_vector[1:3], images = "slice1", pt.size.factor = 1.2) + ggtitle("ZCL 02, B1") + theme(
  plot.title = element_text(face = "bold", size = 30, hjust = 0.5))
ggsave(p, filename = "P1CLB1_cl4_subclusters.pdf", dpi = 300)

p = SpatialPlot(mixed, group.by = "seurat_clusters", cols = col_vector[1:3], images = "slice1.9", pt.size.factor = 1.2) + ggtitle("BnCL 07, A1") + theme(
  plot.title = element_text(face = "bold", size = 30, hjust = 0.5))
ggsave(p, filename = "P4nCLA1_cl4_subclusters.pdf", dpi = 300)
```
